+++ packages/connector-chrome/src/chrome-launcher.ts
@@ -17,7 +17,7 @@ import readFileAsync from 'hint/dist/src/lib/utils/fs/read-file-async';
 import writeFileAsync from 'hint/dist/src/lib/utils/fs/write-file-async';
 
 const debug: debug.IDebugger = d(__filename);
-const lock = promisify(lockfile.lock);
+const lock = promisify(lockfile.lock) as (path: string, options: lockfile.Options) => Promise<void>;
 const unlock = promisify(lockfile.unlock);
 
 export class CDPLauncher extends Launcher {
@@ -25,7 +25,7 @@ export class CDPLauncher extends Launcher {
     private userDataDir: string | boolean;
     private chromeFlags: Array<string>;
 
-    public constructor(options?: LauncherOptions) {
+    public constructor(options: LauncherOptions) {
         super(options);
 
         this.chromeFlags = options && options.flags || ['--no-default-browser-check'];