+++ packages/utils-debugging-protocol-common/src/@types/chrome-remote-interface.d.ts
@@ -1,0 +1,1 @@
+declare module 'chrome-remote-interface';